// dm-frontend/src/app/enums/rarities.enum.ts
export enum RaritiesEnum {
  COMMON = 'COMMON',
  UNCOMMON = 'UNCOMMON',
  RARE = 'RARE',
  VERY_RARE = 'VERY_RARE',
  SUPER_RARE = 'SUPER_RARE',
  VICTORY_RARE = 'VICTORY_RARE',
  DOUBLE_VICTORY_RARE = 'DOUBLE_VICTORY_RARE',
  LEGEND = 'LEGEND',
  FORBIDDEN_LEGEND = 'FORBIDDEN_LEGEND',
  MASTER = 'MASTER',
  SECRET_RARE = 'SECRET_RARE',
}